Remove 'ui_accept' to exit gameplay Scene
First off, thanks for the project skeleton.

This interaction surprised me. In fact, I worked on a project based on this skeleton for a few hours before I dug into
why this strange behavior was happening and realized it's just a line of code in the _process function!

For clarity, I would suggest either (1) removing this logic (this PR) or (2) explaining it in the UI, as you have for the Pause Menu interaction.

Thanks!